json2csv.js
Description: Currently hardcoded to convert a file named "infoList.json" to "infoList.csv"

Currently a file does not exist for infoList.json so it my pull an error when running.

Navigate to the component then "node json2csv.js" to run. This can be used to convert json to csv easily.